+++ src/dcrypt/passblk_mapper.c
@@ -47,8 +47,8 @@ static void putstr(char* buf, char* src, int len)
 static void dm_create(struct part* pt)
 {
 	char* label = pt->label;
-	int nlen = strlen(label);
-	int ret;
+	uint nlen = strlen(label);
+	long ret;
 
 	struct dm_ioctl dmi = {
 		.version = { DM_VERSION_MAJOR, 0, 0 },
@@ -73,13 +73,13 @@ static int dm_single(char* name, uint64_t size, char* targ, char* opts)
 	struct dm_target_spec* dts;
 	char* optstring;
 
-	int nlen = strlen(name);
-	int tlen = strlen(targ);
-	int olen = strlen(opts);
-	int dmilen = sizeof(*dmi);
-	int dtslen = sizeof(*dts);
+	uint nlen = strlen(name);
+	uint tlen = strlen(targ);
+	uint olen = strlen(opts);
+	uint dmilen = sizeof(*dmi);
+	uint dtslen = sizeof(*dts);
 
-	int reqlen = dmilen + dtslen + olen + 1;
+	uint reqlen = dmilen + dtslen + olen + 1;
 	char req[reqlen];
 
 	dmi = (struct dm_ioctl*)(req + 0);
@@ -136,7 +136,7 @@ static int dm_crypt(struct part* pt, char* cipher, void* key, int keylen)
 
 static int dm_suspend(char* name, int flags)
 {
-	int nlen = strlen(name);
+	uint nlen = strlen(name);
 	int ret;
 
 	struct dm_ioctl dmi = {
@@ -159,7 +159,7 @@ static int dm_suspend(char* name, int flags)
 
 static void dm_remove(char* name)
 {
-	int nlen = strlen(name);
+	uint nlen = strlen(name);
 	int ret;
 
 	struct dm_ioctl dmi = {
